requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
marshmallow==3.20.1
python-dotenv==1.0.0
google-generativeai==0.3.2
Werkzeug==2.3.7
bcrypt==4.0.1